test(integ): worker credential file permissions secured
